infrastructure/testdriver/generate_test_report.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver generate_test_report method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
async_test(t => {
  test_driver
    .generate_test_report("Test message.")
    .then((result) => {
        result_html = document.getElementById("json").innerHTML = JSON.stringify(result, undefined, 2);
        assert_equals(result.foo, "bar");
        assert_equals(result.dict["k1"], 5678);
        t.done();
    })
    .catch(t.unreached_func("generate_test_report failed"));
});
</script>

<pre id="json"></pre>
